+++ microlens/src/System/Process/Microlens/CreateProcess.hs
@@ -1,7 +1,6 @@
-{-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE Rank2Types #-}
 -- |
--- Module       : Sysetem.Process.Microlens.CreateProcess
+-- Module       : Sysetem.Process.Lens.CreateProcess
 -- Copyright 	: 2019 Emily Pillmore
 -- License	: BSD
 --
@@ -9,16 +8,28 @@
 -- Stability	: Experimental
 -- Portability	: TypeFamilies, Rank2Types
 --
--- 'CreateProcess' lenses and combinators
+-- This module provides the associated optics and combinators
+-- for working with 'CreateProcess' objects.
+--
+-- Because 'CreateProcess' was created before the `_` prefix record
+-- name convention, some record accessors don't have an apparently
+-- "good" name for their corresponding lens. Those that do not are
+-- post-fixed with `_` i order to denote their lens status. Thankfully,
+-- there are 6 that meet the criteria: 'cmdspec_', 'env_', 'cwd_', 'stdin_',
+-- 'stdout_', and 'stderr_'.
+--
+-- We provide classy variants of what we consider the significant portions
+-- of 'CreateProcess' - namely, the `std_in`, `std_out`, and `std_err` entries.
+--
 --
 module System.Process.Microlens.CreateProcess
 ( -- * Lenses
   cmdspec_
 , cwd_
 , env_
-, stdin
-, stdout
-, stderr
+, stdin_
+, stdout_
+, stderr_
 , closefds
 , creategroup
 , createnewconsole
@@ -62,18 +73,18 @@ env_ = lens env (\t b -> t { env = b })
 
 -- | Lens into the 'std_in' entry of the 'CreateProcess' record
 --
-stdin :: Lens' CreateProcess StdStream
-stdin = lens std_in (\t b -> t { std_in = b })
+stdin_ :: Lens' CreateProcess StdStream
+stdin_ = lens std_in (\t b -> t { std_in = b })
 
 -- | Lens into the 'std_out' entry of the 'CreateProcess' record
 --
-stdout :: Lens' CreateProcess StdStream
-stdout = lens std_out (\t b -> t { std_out = b })
+stdout_ :: Lens' CreateProcess StdStream
+stdout_ = lens std_out (\t b -> t { std_out = b })
 
 -- | Lens into the 'std_err' entry of the 'CreateProcess' record
 --
-stderr :: Lens' CreateProcess StdStream
-stderr = lens std_err (\t b -> t { std_err = b })
+stderr_ :: Lens' CreateProcess StdStream
+stderr_ = lens std_err (\t b -> t { std_err = b })
 
 -- | Lens into the 'close_fds' entry of the 'CreateProcess' record
 --
@@ -132,7 +143,7 @@ instance HasStdin StdStream where
   _Stdin = id
 
 instance HasStdin CreateProcess where
-  _Stdin = stdin
+  _Stdin = stdin_
 
 -- | Classy lens for types with a stdout
 --
@@ -143,7 +154,7 @@ instance HasStdout StdStream where
   _Stdout = id
 
 instance HasStdout CreateProcess where
-  _Stdout = stdout
+  _Stdout = stdout_
 
 -- | Classy lens for types with a stderr
 --
@@ -154,4 +165,4 @@ instance HasStderr StdStream where
   _Stderr = id
 
 instance HasStderr CreateProcess where
-  _Stderr = stderr
+  _Stderr = stderr_