+++ microlens/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Revision history for microlens-process
 
-## 0.1.0.0 -- YYYY-mm-dd
+## 0.1.0.1
+
+Fix the cabal version
+
+## 0.1.0.0
+
+Expose a new module called `System.Process.Microlens.CommandSpec`, exposing
+traversals into the arguments of a raw command.
+
+Added doctests
+
+Improved docs
+
+## 0.0.2.0
+
+Widen bounds, improve docs
+
+## 0.0.1.1
+
+* Correct and widen version bounds
+
+## 0.0.1.0
 
 * First version. Released on an unsuspecting world.